+++ src/drivers/code_hosting_driver.go
@@ -10,7 +10,7 @@ type CodeHostingDriver interface {
 
 	// CanMergePullRequest returns whether or not MergePullRequest should be
 	// called when shipping. If true, also returns the default commit message
-	CanMergePullRequest(branch, parentBranch string) (canMerge bool, defaultCommitMessage string, pullRequestNumber int, err error)
+	CanMergePullRequest(branch, parentBranch string) (canMerge bool, defaultCommitMessage string, pullRequestNumber int64, err error)
 
 	// GetNewPullRequestURL returns the URL of the page
 	// to create a new pull request online
