+++ src/drivers/merge_pull_request_options.go
@@ -3,7 +3,7 @@ package drivers
 // MergePullRequestOptions defines the options to the MergePullRequest function
 type MergePullRequestOptions struct {
 	Branch            string
-	PullRequestNumber int
+	PullRequestNumber int64
 	CommitMessage     string
 	LogRequests       bool
 	ParentBranch      string
